backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.10.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
alembic==1.14.0
psycopg2-binary==2.9.10

# AI/ML Libraries
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
torch==2.5.1
transformers==4.46.0

# Database
asyncpg==0.30.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

